fix: cleanup and move to new file, including test
